Handle window local options correctly
`relativenumber` and `statuscolumn` are window local options, which
makes them difficult to toggle.

So I've made the following changes:

* Window local options (relativenumber and statuscolumn) are only set once
* No more auto commands

After these changes, the plugin is much more reliable. There's one downside
to this new approach: when `enable_in_terminal` is false, the terminal will
display an empty statuscolumn (1-character wide). This feels like a
worthy tradeoff, considering the plugin actually works now, and is even
easier to reason about. I'll be looking for a way to improve this in the future.